1 constant file. move src/constants to src/lib/constants
