+++ src/components/containers/TraceAccordion.js
@@ -1,7 +1,7 @@
 import Fold from './Fold';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
-import {EDITOR_ACTIONS} from '../../constants';
+import {EDITOR_ACTIONS} from '../../lib/constants';
 import {connectTraceToPlot} from '../../lib';
 
 const TraceFold = connectTraceToPlot(Fold);
