+++ src/components/containers/__tests__/Layout-test.js
@@ -2,7 +2,7 @@ import {Numeric} from '../../fields';
 import {Fold, Panel, Section} from '..';
 import NumericInput from '../../widgets/NumericInputStatefulWrapper';
 import React from 'react';
-import {EDITOR_ACTIONS} from '../../../constants';
+import {EDITOR_ACTIONS} from '../../../lib/constants';
 import {TestEditor, fixtures} from '../../../lib/test-utils';
 import {connectLayoutToPlot} from '../../../lib';
 import {mount} from 'enzyme';
