+++ src/lib/connectLayoutToPlot.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import nestedProperty from 'plotly.js/src/lib/nested_property';
 import {getDisplayName, unpackPlotProps} from '../lib';
-import {EDITOR_ACTIONS} from '../constants';
+import {EDITOR_ACTIONS} from './constants';
 
 export const getLayoutContext = context => {
   const {layout, fullLayout, plotly, onUpdate} = context;
